+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.22.1 (2025-09-11)
+
+- Add optional `routing_key` parameter to `#published_messages` in `Ears::Testing::TestHelper`
+- Add `#last_published_payload` to `Ears::Testing::TestHelper`
+
 ## 0.22.0 (2025-09-11)
 
 - Support publisher confirms in `Ears::Publisher`